+++ app/Events/NewRequestRequested.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Events;
+
+use Illuminate\Foundation\Events\Dispatchable;
+
+class NewRequestRequested
+{
+    use Dispatchable;
+
+    public function __construct(mixed ...$args) {}
+}